+++ mariadb-practices/jdbc-practices/src/main/java/bookmall/vo/OrderBookVo.java
@@ -1,42 +1,42 @@
 package bookmall.vo;
 
 public class OrderBookVo {
-    private int no;
-    private int orderNo;
-    private int bookNo;
+    private long no;
+    private long orderNo;
+    private long bookNo;
     private int quantity;
     private int price;
 
     public OrderBookVo() {}
 
-    public OrderBookVo(int orderNo, int bookNo, int quantity, int price) {
+    public OrderBookVo(long orderNo, long bookNo, int quantity, int price) {
         this.orderNo = orderNo;
         this.bookNo = bookNo;
         this.quantity = quantity;
         this.price = price;
     }
 
-    public int getNo() {
+    public long getNo() {
         return no;
     }
 
-    public void setNo(int no) {
+    public void setNo(long no) {
         this.no = no;
     }
 
-    public int getOrderNo() {
+    public long getOrderNo() {
         return orderNo;
     }
 
-    public void setOrderNo(int orderNo) {
+    public void setOrderNo(long orderNo) {
         this.orderNo = orderNo;
     }
 
-    public int getBookNo() {
+    public long getBookNo() {
         return bookNo;
     }
 
-    public void setBookNo(int bookNo) {
+    public void setBookNo(long bookNo) {
         this.bookNo = bookNo;
     }
